% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASMR-package.R
\docType{package}
\name{ASMR-package}
\alias{ASMR}
\alias{ASMR-package}
\title{ASMR: Analysis of Selection-biased Matched data in R}
\description{
R package \bold{ASMR} is a package aimed at xxx.
}
\references{
Lee, S., Sim, H., and Lee, W. (2025). Developing statistical methods for selection-biased matched data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seungjae2525/ASMR}
  \item Report bugs at \url{https://github.com/seungjae2525/ASMR/issues}
}

}
\author{
Seungjae Lee \email{seungjae2525@gmail.com}, Hyunman Sim \email{smisk77@gmail.com}, and Woojoo Lee \email{lwj221@gmail.com}
}
\keyword{package}
