% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.asmr.R
\name{summary.asmr}
\alias{summary.asmr}
\title{Summary for \code{asmr} objects}
\usage{
\method{summary}{asmr}(object, digits = max(1L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{An object for class \code{asmr}.}

\item{digits}{Number of significant digits. Default: \code{max(1L, getOption("digits") - 3L)}.}

\item{...}{Further arguments (currently not used).}
}
\value{
No return value, called for side effects.
}
\description{
Summary the results for object of class \code{asmr}.
}
\details{
Summary the results for object of class \code{asmr}.
From the conditional logistic regression results, the odds ratio, confidence interval of the odds ratio, and \emph{P} value are reported.
}
\examples{
## Load example data
data("backPain", package = "gnm")
## Re-express as count data
backPainLong <- gnm::expandCategorical(backPain, "pain")

}
\seealso{
\code{\link[ASMR]{asmr}}, \code{\link[ASMR]{print.asmr}}, \code{\link[base]{summary}}
}
\keyword{summary}
