% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.asmr.R
\name{print.asmr}
\alias{print.asmr}
\title{Print for \code{asmr} objects}
\usage{
\method{print}{asmr}(x, ...)
}
\arguments{
\item{x}{An object for class \code{asmr}.}

\item{...}{Further arguments (currently not used).}
}
\value{
No return value, called for side effects.
}
\description{
Print the results for object of class \code{asmr}.
}
\details{
Print the results for object of class \code{asmr}.
From the conditional logistic regression results, \dQuote{Estimate} corresponds to the log odds ratio and \dQuote{SE} corresponds to the standard error of the log odds ratio.
}
\examples{
## Load example data
data("backPain", package = "gnm")
## Re-express as count data
backPainLong <- gnm::expandCategorical(backPain, "pain")

}
\seealso{
\code{\link[ASMR]{asmr}}, \code{\link[ASMR]{summary.asmr}}, \code{\link[base]{print}}
}
\keyword{print}
